requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-SQLAlchemy==1.0
Flask-uWSGI-WebSocket==0.2.5
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.6
SQLAlchemy-JSONAPI==0.2
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.13
itsdangerous==0.24
kombu==3.0.21
psycopg2==2.5.3
pyinotify==0.9.4
pytaglib==0.4.1
python-magic==0.4.6
pytz==2014.4
redis==2.10.2
uWSGI==2.0.6
webassets==0.10.1
wsgiref==0.1.2
